requirements.txt
zendriver==0.15.2
pyYAML==6.0.3
stackprinter==0.2.12
loguru==0.7.3
regex-string-generator==0.1.0